Separate JRE GC log file only supported on Oracle JRE
While we seem to be able to enable GC logging on a JREs, only
Oracle supports writing those to a separate file. The IBM JRE
does not support this, nor does the OpenJDK. So, we restrict
the logging to only x86_64 platforms (which is all we deploy
the Oracle JRE on currently).